+++ src/main/java/corecord/dev/domain/analysis/dto/response/AnalysisResponse.java
@@ -1,6 +1,6 @@
 package corecord.dev.domain.analysis.dto.response;
 
-import corecord.dev.domain.analysis.constant.Keyword;
+import corecord.dev.domain.Ability.dto.response.AbilityResponse;
 import corecord.dev.domain.record.constant.RecordType;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -10,15 +10,6 @@ import lombok.Getter;
 import java.util.List;
 
 public class AnalysisResponse {
-
-    @Data
-    @Builder
-    @Getter
-    @AllArgsConstructor
-    public static class AbilityDto {
-        private String keyword;
-        private String content;
-    }
 
     @Builder
     @Getter
@@ -31,37 +22,9 @@ public class AnalysisResponse {
         private RecordType recordType;
         private String recordTitle;
         private String recordContent;
-        private List<AbilityDto> abilityDtoList;
+        private List<AbilityResponse.AbilityDto> abilityDtoList;
         private String comment;
         private String createdAt;
     }
 
-    @Builder
-    @Getter
-    @AllArgsConstructor
-    @Data
-    public static class KeywordListDto {
-        private List<String> keywordList;
-    }
-
-    @Data
-    public static class KeywordStateDto {
-        private String keyword;
-        private Long count;
-        private Long percent;
-
-        public KeywordStateDto(Keyword keyword, Long count, Double percent) {
-            this.keyword = keyword.getValue();
-            this.count = count;
-            this.percent = Math.round(percent);
-        }
-    }
-
-    @Builder
-    @Getter
-    @AllArgsConstructor
-    @Data
-    public static class GraphDto {
-        List<KeywordStateDto> keywordGraph;
-    }
 }
